SCC-2203 Add rubocop and associated linting fixes
This implements rubocop to lint the repo and includes all the changes flagged by running that. The `.rubocop.yml` file itself is somewhat customized but generally contains standards. This also adds it as a step in the travis integration to enforce standards before a commit/pr passes the test suite.