SCC-2204 Improve performance via batched querying
This improves the performance of the check-in card poller by limiting the number of records retrieved in a single request. Ocassional timeouts were being received and this resolves that by ensuring that the requests are always of a reasonable size. This involved a slight reorganization of the logic and updating some configuration