NOREF Ignore Duplicate rows
For not fully understood reasons the current process can produce duplicate rows in the SQLITE database. This implements a `UNIQUE` constraint on the `id` column to ensure that each row is distinct and adds an `OR IGNORE` clause to the `INSERT` query to quietly skip any rows that violate that constraint.